feat: avoid killing the server on bad subscription
If the query subscription had something that isn't on the schema,
for example an nonexistent branch, then the server would throw an
error and die.
This commit catches the error, logs with dlt, cancels the
subscription and sends a response containing the error to the
client. The server will also stay alive.
When calling the resolve or subscribe methods on the resolvable
schema object, errors will be caught and reported to dlt before
thrown.
This is necessary since we can't see the errors when we don't start
the server using command line or when the std::terminate causes system
crash, unless the dlt server is reporting.

Change-Id: I53e812b3dd33204f5f702ff6c42f749095c9189d