fix: remove absolute INCLUDE path caused by dlt
The `pkg_check_modules(DLT REQUIRED automotive-dlt)` creates a
variable DLT_INCLUDE_DIRS containing the include directory for dlt.

A string with two paths is created to support "#include <dlt.h>"
and `#include <dlt/dlt.h>`. This first is the 'legacy form'.
The paths are separated by a semicolon.

The $<BUILD_INTERFACE:${DLT_INCLUDE_DIRS}> expanded in two strings,
causing the second one to include a absolute path to the dlt
includes.

To solve that, we wrap the `$<BUILD_INTERFACE:${DLT_INCLUDE_DIRS}>`
in quotes, making the two paths to be under the BUILD_INTERFACE
directive.

Change-Id: I6bf26924ca9b662f98742c91107ec0d17afb1aa2